Rewrite MiniTest tests as Rspec specs
Rspec is very popular amongst Ruby developers and used in more gems than
MiniTest. It also offers more features down the line and I noticed that
it was 20% faster.

It's easier to switch now than later when we got more tests.
