+++ src/org/datafoodconsortium/connector/codegen/ruby/static/.github/workflows/ruby.yml
@@ -25,4 +25,4 @@ jobs:
         ruby-version: ${{ matrix.ruby-version }}
         bundler-cache: true # runs 'bundle install' and caches installed gems automatically
     - name: Run tests
-      run: bundle exec rake
+      run: bundle exec rspec
