+++ src/org/datafoodconsortium/connector/codegen/ruby/static/spec/supplied_product_spec.rb
@@ -1,17 +1,12 @@
-require 'connector'
-require "minitest/autorun"
-require 'datafoodconsortium/connector'
-
-class SuppliedProductTest < Minitest::Test
-
-  def test_export_empty
+describe DataFoodConsortium::Connector::SuppliedProduct do
+  it "can be empty when exported" do
     sp = DataFoodConsortium::Connector::SuppliedProduct.new("https://myplatform.com/sp")
     actual = connector.export(sp)
     expected = '{"@context":"https://www.datafoodconsortium.org","@id":"https://myplatform.com/sp","@type":"dfc-b:SuppliedProduct","dfc-b:name":"","dfc-b:description":"","dfc-b:alcoholPercentage":0.0,"dfc-b:lifetime":"","dfc-b:usageOrStorageCondition":"","dfc-b:totalTheoreticalStock":0.0}'
-    assert_equal expected, actual
+    expect(actual).to eq expected
   end
 
-  def test_export_complete
+  it "contains all fields when exported" do
     quantity = DataFoodConsortium::Connector::QuantitativeValue.new(
       unit: connector.MEASURES.KILOGRAM,
       value: 1.2
@@ -57,7 +52,6 @@ class SuppliedProductTest < Minitest::Test
 
     actual = connector.export(sp)
     expected = '{"@context":"https://www.datafoodconsortium.org","@id":"https://myplatform.com/sp","@type":"dfc-b:SuppliedProduct","dfc-b:name":"name","dfc-b:description":"description","dfc-b:hasType":"dfc-pt:artichoke","dfc-b:hasQuantity":{"@type":"dfc-b:QuantitativeValue","dfc-b:hasUnit":"dfc-m:Kilogram","dfc-b:value":1.2},"dfc-b:alcoholPercentage":2.6,"dfc-b:lifetime":"lifetime","dfc-b:hasClaim":["dfc-f:EnergyFree","dfc-f:FatFree","dfc-f:HighFibre"],"dfc-b:usageOrStorageCondition":"usageOrStorageConditions","dfc-b:hasNutrientCharacteristic":{"@type":"dfc-b:NutrientCharacteristic","dfc-b:hasUnit":"dfc-m:Gram","dfc-b:value":8.47,"dfc-b:hasNutrientDimension":"dfc-m:Calcium"},"dfc-b:hasPhysicalCharacteristic":{"@type":"dfc-b:PhysicalCharacteristic","dfc-b:hasUnit":"dfc-m:Kilogram","dfc-b:value":3.25,"dfc-b:hasPhysicalDimension":"dfc-m:Weight"},"dfc-b:hasGeographicalOrigin":"dfc-f:Normandy","dfc-b:hasCertification":"dfc-f:Demeter","dfc-b:hasNatureOrigin":"https://github.com/datafoodconsortium/taxonomies/releases/latest/download/facets.rdf#PlantOrigin","dfc-b:hasPartOrigin":"https://github.com/datafoodconsortium/taxonomies/releases/latest/download/facets.rdf#PlantOrigin","dfc-b:totalTheoreticalStock":123}'
-    assert_equal expected, actual
+    expect(actual).to eq expected
   end
-
 end
