+++ src/org/datafoodconsortium/connector/codegen/ruby/static/spec/support/productTypes.json
